agentlend-ai/backend/requirements.txt
# AgentLend AI — Backend Dependencies

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Validation & serialization
pydantic==2.10.3
pydantic-settings==2.7.0

# Environment configuration
python-dotenv==1.0.1

# HTTP client
requests==2.32.3
httpx==0.28.1

# Google Gemini LLM
google-generativeai==0.8.3

# Ethereum / Web3
web3==7.6.0

# Background task scheduling
apscheduler==3.10.4

# Logging
structlog==24.4.0
